3d mesh rendering

+++ src/draw.h
@@ -5,6 +5,6 @@ namespace draw
 {
 
 void DrawTriangle(size_t size, Vertex_ *vertices);
-void DrawIndexedTriangle(size_t size, Vertex_* vertices, size_t index_count, uint16_t* index_array);
+void DrawIndexedTriangle(size_t size, Vertex_* vertices, size_t index_count, uint32_t* index_array);
 
 } //namespace draw
